bookstore-microservice/staff-service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.2
psycopg2-binary>=2.9
requests>=2.31
gunicorn>=21.2
